SpeakerFX: Improve output detection logic
Instead of detecting based on the first device, detect based on the current
active output device.

This fixes the EQ not working properly if the user has multiple outputs
connected and switches them manually without disconnecting (such as using
the device picker in the volume bar, or the Now Playing tile).

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>